first pass at interactive envelope parameters. i hear rice crispies but im not sure if thats in my envelope math. my envelopes have linear movement, i believe they should be exponential

+++ include/gain.h
@@ -5,7 +5,7 @@ class Gain
 {
 public:
 	Gain();
-	void setGaindB(int gaindB);
+	void setGaindB(float gaindB);
 	float apply(float samp);
 
 private:
